task-queue-backend/pyproject.toml
[project]
name = "task-queue-backend"
version = "0.1.0"
description = "TIRA Task Queue Backend — FastAPI + Celery client REST API"
requires-python = ">=3.13"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "celery[redis]",
    "pydantic-settings",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/task_queue_backend"]
